MULE-13711: JMS Publisher always joins the current TX, if one exists, regardless the chosen transactional Action
